core/src/main/java/net/cactusthorn/routing/convert/IntegerConverter.java
package net.cactusthorn.routing.convert;

import net.cactusthorn.routing.RequestData;

public class IntegerConverter implements Converter {

    @Override //
    public Integer convert(RequestData requestData, Class<?> type, String value) {
        if (value == null || value.trim().isEmpty()) {
            return null;
        }
        return Integer.valueOf(value);
    }

    @Override //
    public Object convert(RequestData requestData, Class<?> type, String[] value) throws ConverterException {
        if (value == null) {
            return null;
        }

        Integer[] array = new Integer[value.length];
        for (int i = 0; i < value.length; i++) {
            array[i] = convert(requestData, type, value[i]);
        }
        return array;
    }
}
